ref: moved util classes to util package
