+++ src/main/java/se2/server/hanabi/game/HintType.java
@@ -1,4 +1,4 @@
-package se2.server.hanabi.gamemanager;
+package se2.server.hanabi.game;
 
 public enum HintType {
     COLOR,
