+++ src/main/java/se2/server/hanabi/game/actions/DiscardCardAction.java
@@ -1,9 +1,8 @@
-package se2.server.hanabi.gamemanager.actions;
+package se2.server.hanabi.game.actions;
 
-import se2.server.hanabi.gamemanager.GameManager;
+import se2.server.hanabi.game.GameManager;
 import se2.server.hanabi.model.Card;
-import se2.server.hanabi.rules.GameRules;
-import se2.server.hanabi.gamemanager.ActionResult;
+import se2.server.hanabi.util.ActionResult;
 
 import java.util.List;
