+++ src/test/java/se2/server/hanabi/GameManagerTest.java
@@ -3,11 +3,11 @@ package se2.server.hanabi;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import se2.server.hanabi.api.GameStatus;
-import se2.server.hanabi.gamemanager.ActionResult;
-import se2.server.hanabi.gamemanager.GameManager;
-import se2.server.hanabi.gamemanager.HintType;
+import se2.server.hanabi.util.ActionResult;
+import se2.server.hanabi.game.GameManager;
+import se2.server.hanabi.game.HintType;
 import se2.server.hanabi.model.Card;
-import se2.server.hanabi.rules.GameRules;
+import se2.server.hanabi.util.GameRules;
 
 import java.util.Arrays;
 import java.util.List;
